+++ examples/basic/client.js
@@ -1,0 +1,19 @@
+'use strict'
+
+const Hemera = require('./../../packages/hemera')
+const nats = require('nats').connect()
+
+const hemera = new Hemera(nats, {
+  logLevel: 'info'
+})
+
+hemera.ready(() => {
+  hemera.act({
+    topic: 'math',
+    cmd: 'add',
+    a: 1,
+    b: 2
+  }, function (err, resp) {
+    this.log.info(resp, 'Result')
+  })
+})