+++ examples/basic/server.js
@@ -1,0 +1,20 @@
+'use strict'
+
+const Hemera = require('./../../packages/hemera')
+const nats = require('nats').connect()
+
+const hemera = new Hemera(nats, {
+  logLevel: 'info'
+})
+
+hemera.ready(() => {
+  /**
+   * Your Implementations
+   */
+  hemera.add({
+    topic: 'math',
+    cmd: 'add'
+  }, function (req, cb) {
+    cb(null, req.a + req.b)
+  })
+})