+++ packages/hemera-mongo-store/index.js
@@ -15,7 +15,9 @@ exports.plugin = function hemeraMongoStore (options, next) {
   }
 
   Mongodb.MongoClient.connect(options.mongo.url, options.mongos.options, function (err, db) {
-    if (err) throw err
+    if (err) {
+      return hemera.emit('error', err)
+    }
 
     hemera.expose('db', db)
     hemera.expose('mongodb', Mongodb)